requirements.txt
Flask==2.2.2
gunicorn== 21.2.0
itsdangerous==2.0.1
Jinja2== 3.1.2
MarkupSafe== 2.1.1
Werkzeug== 2.2.3
numpy>= 1.24.3
scipy>= 1.11.1
scikit-learn== 1.3.0
matplotlib>= 3.7.2
pandas>= 2.0.3
mysql-connector-python==8.2.0
nltk==3.8.1
joblib== 1.2.0
pycaret==3.2.0
certifi==2023.11.17
streamlit

